pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-session>=0.8.0",
    "pillow>=11.3.0",
    "requests>=2.32.5",
    "stability-sdk>=0.8.6",
    "werkzeug>=3.1.3",
]
